- Updated Newtonsoft.Json to 10.0.3 - Added Elastic Email support. - Refactoring.

+++ src/Frapid.Web/Areas/SendGridMail/Processor.cs
@@ -118,7 +118,7 @@ namespace SendGridMail
             }
             finally
             {
-                if (deleteAttachmentes)
+                if (deleteAttachmentes && email.Status == Status.Completed)
                 {
                     FileHelper.DeleteFiles(attachments);
                 }
